chore: Rename pl.dependencies to pl._dependencies
