+++ py-polars/polars/_typing.py
@@ -24,11 +24,11 @@ if TYPE_CHECKING:
     from sqlalchemy.orm import Session
 
     from polars import DataFrame, Expr, LazyFrame, Series
+    from polars._dependencies import numpy as np
+    from polars._dependencies import pandas as pd
+    from polars._dependencies import pyarrow as pa
+    from polars._dependencies import torch
     from polars.datatypes import DataType, DataTypeClass, IntegerType, TemporalType
-    from polars.dependencies import numpy as np
-    from polars.dependencies import pandas as pd
-    from polars.dependencies import pyarrow as pa
-    from polars.dependencies import torch
     from polars.lazyframe.engine_config import GPUEngine
     from polars.selectors import Selector
