+++ py-polars/polars/_utils/async_.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 from collections.abc import Awaitable
 from typing import TYPE_CHECKING, Any, Generic, TypeVar
 
+from polars._dependencies import _GEVENT_AVAILABLE
 from polars._utils.wrap import wrap_df
-from polars.dependencies import _GEVENT_AVAILABLE
 
 if TYPE_CHECKING:
     from asyncio.futures import Future
