+++ py-polars/polars/_utils/getitem.py
@@ -5,6 +5,8 @@ from typing import TYPE_CHECKING, Any, NoReturn, overload
 
 import polars._reexport as pl
 import polars.functions as F
+from polars._dependencies import _check_for_numpy
+from polars._dependencies import numpy as np
 from polars._utils.constants import U32_MAX
 from polars._utils.slice import PolarsSlice
 from polars._utils.various import qualified_type_name, range_to_slice
@@ -18,8 +20,6 @@ from polars.datatypes.classes import (
     UInt32,
     UInt64,
 )
-from polars.dependencies import _check_for_numpy
-from polars.dependencies import numpy as np
 from polars.meta.index_type import get_index_type
 
 if TYPE_CHECKING:
