+++ py-polars/polars/_utils/various.py
@@ -28,6 +28,8 @@ from typing import (
 
 import polars as pl
 from polars import functions as F
+from polars._dependencies import _check_for_numpy, import_optional, subprocess
+from polars._dependencies import numpy as np
 from polars.datatypes import (
     Boolean,
     Date,
@@ -39,8 +41,6 @@ from polars.datatypes import (
     Time,
 )
 from polars.datatypes.group import FLOAT_DTYPES, INTEGER_DTYPES
-from polars.dependencies import _check_for_numpy, import_optional, subprocess
-from polars.dependencies import numpy as np
 
 if TYPE_CHECKING:
     from collections.abc import Iterator, MutableMapping, Reversible
