+++ py-polars/polars/dataframe/frame.py
@@ -30,6 +30,23 @@ from typing import (
 
 import polars._reexport as pl
 from polars import functions as F
+from polars._dependencies import (
+    _ALTAIR_AVAILABLE,
+    _GREAT_TABLES_AVAILABLE,
+    _PANDAS_AVAILABLE,
+    _PYARROW_AVAILABLE,
+    _check_for_numpy,
+    _check_for_pandas,
+    _check_for_pyarrow,
+    _check_for_torch,
+    altair,
+    great_tables,
+    import_optional,
+    torch,
+)
+from polars._dependencies import numpy as np
+from polars._dependencies import pandas as pd
+from polars._dependencies import pyarrow as pa
 from polars._typing import DbWriteMode, JaxExportType, TorchExportType
 from polars._utils.construction import (
     arrow_to_pydf,
@@ -82,23 +99,6 @@ from polars.datatypes import (
     UInt64,
 )
 from polars.datatypes.group import INTEGER_DTYPES
-from polars.dependencies import (
-    _ALTAIR_AVAILABLE,
-    _GREAT_TABLES_AVAILABLE,
-    _PANDAS_AVAILABLE,
-    _PYARROW_AVAILABLE,
-    _check_for_numpy,
-    _check_for_pandas,
-    _check_for_pyarrow,
-    _check_for_torch,
-    altair,
-    great_tables,
-    import_optional,
-    torch,
-)
-from polars.dependencies import numpy as np
-from polars.dependencies import pandas as pd
-from polars.dependencies import pyarrow as pa
 from polars.exceptions import (
     ColumnNotFoundError,
     InvalidOperationError,
