+++ py-polars/polars/io/_utils.py
@@ -7,12 +7,12 @@ from io import BytesIO, StringIO
 from pathlib import Path
 from typing import IO, TYPE_CHECKING, Any, overload
 
+from polars._dependencies import _FSSPEC_AVAILABLE, fsspec
 from polars._utils.various import (
     is_int_sequence,
     is_str_sequence,
     normalize_filepath,
 )
-from polars.dependencies import _FSSPEC_AVAILABLE, fsspec
 from polars.exceptions import NoDataError
 
 if TYPE_CHECKING:
