+++ py-polars/polars/io/database/functions.py
@@ -3,10 +3,10 @@ from __future__ import annotations
 import re
 from typing import TYPE_CHECKING, Any, Literal, overload
 
+from polars._dependencies import import_optional
 from polars._utils.unstable import issue_unstable_warning
 from polars._utils.various import qualified_type_name
 from polars.datatypes import N_INFER_DEFAULT
-from polars.dependencies import import_optional
 from polars.io.database._cursor_proxies import ODBCCursorProxy
 from polars.io.database._executor import ConnectionExecutor
