+++ py-polars/polars/io/spreadsheet/functions.py
@@ -14,6 +14,7 @@ from typing import IO, TYPE_CHECKING, Any, Callable, NoReturn, overload
 import polars._reexport as pl
 from polars import from_arrow
 from polars import functions as F
+from polars._dependencies import import_optional
 from polars._utils.deprecation import (
     deprecate_renamed_parameter,
     issue_deprecation_warning,
@@ -32,7 +33,6 @@ from polars.datatypes import (
     UInt8,
 )
 from polars.datatypes.group import FLOAT_DTYPES, INTEGER_DTYPES, NUMERIC_DTYPES
-from polars.dependencies import import_optional
 from polars.exceptions import (
     ModuleUpgradeRequiredError,
     NoDataError,
