+++ py-polars/polars/ml/utilities.py
@@ -1,9 +1,9 @@
 from typing import Any
 
 from polars import DataFrame
+from polars._dependencies import numpy as np
 from polars._typing import IndexOrder
 from polars.datatypes import Array, List
-from polars.dependencies import numpy as np
 
 
 def frame_to_numpy(
