+++ py-polars/pyproject.toml
@@ -210,7 +210,7 @@ ignore = [
 allowed-confusables = ["µ"]
 
 [tool.ruff.lint.per-file-ignores]
-"dependencies.py" = ["ICN001"]
+"_dependencies.py" = ["ICN001"]
 "tests/**/*.py" = ["D100", "D102", "D103", "B018", "FBT001"]
 
 [tool.ruff.lint.pycodestyle]
